Reorganize solution into several projects to modularize logic and presentation.
