--- ConsoleSolution/ConsoleSolution.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.6.33829.357
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConsoleSolution", "ConsoleSolution\ConsoleSolution.csproj", "{179C1B12-40CA-42EA-AE23-A4A0C0D52C6D}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SolutionUnitTests", "SolutionUnitTests\SolutionUnitTests.csproj", "{48C3C33D-6F6F-4D20-B595-EDBF89EAB80D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{179C1B12-40CA-42EA-AE23-A4A0C0D52C6D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{179C1B12-40CA-42EA-AE23-A4A0C0D52C6D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{179C1B12-40CA-42EA-AE23-A4A0C0D52C6D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{179C1B12-40CA-42EA-AE23-A4A0C0D52C6D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{48C3C33D-6F6F-4D20-B595-EDBF89EAB80D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{48C3C33D-6F6F-4D20-B595-EDBF89EAB80D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{48C3C33D-6F6F-4D20-B595-EDBF89EAB80D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{48C3C33D-6F6F-4D20-B595-EDBF89EAB80D}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A6E0C327-9EE0-431F-99B0-3FE9F37B1976}
-	EndGlobalSection
-EndGlobal
