--- ConsoleSolution/SolutionUnitTests/Usings.cs
@@ -1,9 +1,0 @@
-global using Microsoft.VisualStudio.TestTools.UnitTesting;
-global using ConsoleSolution.Models.Json;
-global using Newtonsoft.Json;
-global using ConsoleSolution.Interfaces;
-global using ConsoleSolution.Objects;
-global using ConsoleSolution.Objects.AnswerProviders;
-
-// Credit Moq. https://github.com/devlooped/moq
-global using Moq;
