+++ TechnicalInterviewSolution/AutomatedTests/Usings.cs
@@ -1,0 +1,12 @@
+global using Microsoft.VisualStudio.TestTools.UnitTesting;
+global using Backend.Models.Json;
+global using Backend.Interfaces;
+global using Backend.Objects;
+global using Backend.Objects.AnswerProviders;
+global using Backend.Models.Sql;
+
+// Credit Moq. https://github.com/devlooped/moq
+global using Moq;
+
+// Credit Newtonsoft. https://www.newtonsoft.com/json
+global using Newtonsoft.Json;
