+++ TechnicalInterviewSolution/Backend/Models/Sql/AnswerRecord.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Backend.Models.Sql
+{
+    public class AnswerRecord
+    {
+        public int AnswerRecordID { get; set; }
+        public DateTime AnswerDate { get; set; }
+        public string FileName { get; set; }
+        public string Question { get; set; }
+        public string Answer { get; set; }
+    }
+}
